function/requirements.txt
azure-functions
azure-identity
requests
